+++ lib/canvas/canvas-core/options/create-options.ts
@@ -2,11 +2,11 @@ import { standardCenterFn } from "@/center-fn";
 import { CoreOptions } from "./core-options";
 import { resolveEdgeShapeFactory } from "./resolve-edge-shape-factory";
 import { resolvePriority } from "./resolve-priority";
-import { GraphStoreControllerOptions } from "@/graph-store-controller";
+import { GraphStoreControllerDefaults } from "@/graph-store-controller";
 
 export const createOptions: (
   apiOptions: CoreOptions | undefined,
-) => GraphStoreControllerOptions = (apiOptions: CoreOptions | undefined) => {
+) => GraphStoreControllerDefaults = (apiOptions: CoreOptions | undefined) => {
   const priorities = resolvePriority(
     apiOptions?.nodes?.priority,
     apiOptions?.edges?.priority,
