+++ lib/canvas/core-canvas/core-canvas.ts
@@ -1,5 +1,3 @@
-import { CoreOptions, createOptions } from "./options";
-import { GraphStore } from "@/graph-store";
 import {
   PublicViewportTransformer,
   ViewportTransformer,
@@ -13,11 +11,9 @@ import { MarkPortRequest } from "../mark-port-request";
 import { UpdatePortRequest } from "../update-port-request";
 import { PatchMatrixRequest } from "../patch-matrix-request";
 import { HtmlController } from "@/html-controller";
-import {
-  GraphStoreController,
-  GraphStoreControllerEvents,
-} from "@/graph-store-controller";
+import { GraphStoreController } from "@/graph-store-controller";
 import { PublicGraphStore } from "@/public-graph-store";
+import { DiContainer } from "./di-container";
 
 /**
  * Provides low level API for acting on graph
@@ -29,75 +25,16 @@ export class CoreCanvas implements Canvas {
 
   private readonly internalTransformation: ViewportTransformer;
 
-  private readonly internalModel: GraphStore;
-
   private readonly graphStoreController: GraphStoreController;
 
   private readonly htmlController: HtmlController;
 
-  private readonly events: GraphStoreControllerEvents = {
-    onAfterNodeAdded: (nodeId): void => {
-      this.htmlController.attachNode(nodeId);
-    },
-    onAfterEdgeAdded: (edgeId): void => {
-      this.htmlController.attachEdge(edgeId);
-    },
-    onAfterEdgeShapeUpdated: (edgeId): void => {
-      this.htmlController.updateEdgeShape(edgeId);
-    },
-    onAfterEdgePriorityUpdated: (edgeId): void => {
-      this.htmlController.updateEdgePriority(edgeId);
-    },
-    onAfterEdgeUpdated: (edgeId): void => {
-      this.htmlController.renderEdge(edgeId);
-    },
-    onAfterPortUpdated: (portId): void => {
-      const edges = this.internalModel.getPortAdjacentEdgeIds(portId);
-
-      edges.forEach((edge) => {
-        this.htmlController.renderEdge(edge);
-      });
-    },
-    onAfterNodePriorityUpdated: (nodeId): void => {
-      this.htmlController.updateNodePriority(nodeId);
-    },
-    onAfterNodeUpdated: (nodeId): void => {
-      this.htmlController.updateNodeCoordinates(nodeId);
-      const edges = this.internalModel.getNodeAdjacentEdgeIds(nodeId);
-
-      edges.forEach((edge) => {
-        this.htmlController.renderEdge(edge);
-      });
-    },
-    onBeforeEdgeRemoved: (edgeId): void => {
-      this.htmlController.detachEdge(edgeId);
-    },
-    onBeforeNodeRemoved: (nodeId): void => {
-      this.htmlController.detachNode(nodeId);
-    },
-  };
-
-  public constructor(apiOptions?: CoreOptions) {
-    this.internalModel = new GraphStore();
-    this.model = new PublicGraphStore(this.internalModel);
-
-    this.internalTransformation = new ViewportTransformer(() => {
-      this.htmlController.applyTransform();
-    });
-    this.transformation = new PublicViewportTransformer(
-      this.internalTransformation,
-    );
-
-    this.htmlController = new HtmlController(
-      this.internalModel,
-      this.internalTransformation,
-    );
-
-    this.graphStoreController = new GraphStoreController(
-      this.internalModel,
-      createOptions(apiOptions),
-      this.events,
-    );
+  public constructor(di: DiContainer) {
+    this.model = di.model;
+    this.internalTransformation = di.internalTransformation;
+    this.transformation = di.transformation;
+    this.htmlController = di.htmlController;
+    this.graphStoreController = di.graphStoreController;
   }
 
   public attach(element: HTMLElement): CoreCanvas {
