+++ lib/canvas/core-canvas/index.ts
@@ -1,2 +1,3 @@
 export { CoreCanvas } from "./core-canvas";
 export type { CoreOptions, EdgeShape } from "./options";
+export { DiContainer } from "./di-container";
