+++ lib/canvas/core-canvas/options/index.ts
@@ -1,4 +1,4 @@
-export { createOptions } from "./create-options";
+export { createDefaults } from "./create-defaults";
 export type { CoreOptions } from "./core-options";
 export type { EdgeShape } from "./edge-shape";
 export type { Priority } from "./priority";
