+++ lib/graph-store-controller/graph-store-controller-defaults.ts
@@ -2,7 +2,7 @@ import { CenterFn } from "@/center-fn";
 import { EdgeShapeFactory } from "./edge-shape-factory";
 import { PriorityFn } from "@/priority";
 
-export interface GraphStoreControllerOptions {
+export interface GraphStoreControllerDefaults {
   readonly nodes: {
     readonly centerFn: CenterFn;
     readonly priorityFn: PriorityFn;
